// apps/web/src/lib/service-clients/service-properties/generated/schemas/propertyOptionResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * properties_service
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyOptionResponseColor } from './propertyOptionResponseColor';
import type { PropertyOptionValue } from './propertyOptionValue';

/**
 * Property option response (API representation).
 */
export interface PropertyOptionResponse {
  color?: PropertyOptionResponseColor;
  displayOrder: number;
  id: string;
  propertyDefinitionId: string;
  value: PropertyOptionValue;
}